+++ jpegwrapper/jpeg_mem.cpp
@@ -80,7 +80,7 @@ void save_jpeg_mem(const fs_image_matrix &matrix,
 std::string jwp_save_jpeg_mem_as_string(const fs_image_matrix & matrix, const unsigned int quality) {
 	std::string out;
 	save_jpeg_mem(matrix,
-		[&](const uint8_t *img, unsigned long size) {
+		[&](const uint8_t *img, size_t size) {
 		out = std::string((char*)img, size);
 	}, quality);
 	return out;
@@ -88,7 +88,7 @@ std::string jwp_save_jpeg_mem_as_string(const fs_image_matrix & matrix, const un
 std::vector<uint8_t> jwp_save_jpeg_mem_as_vector(const fs_image_matrix & matrix, const unsigned int quality) {
 	std::vector<uint8_t> out;
 	save_jpeg_mem(matrix,
-		[&](const uint8_t *img, unsigned long size) {
+		[&](const uint8_t *img, size_t size) {
 		out = std::vector<uint8_t>(img, img + size);
 	}, quality);
 	return out;
@@ -105,7 +105,7 @@ void save_jpeg_gray_mem(const fs_image_matrix &matrix,
 std::string jwp_save_jpeg_gray_mem_as_string(const fs_image_matrix & matrix, const unsigned int quality) {
 	std::string out;
 	save_jpeg_gray_mem(matrix,
-		[&](const uint8_t *img, unsigned long size) {
+		[&](const uint8_t *img, size_t size) {
 		out = std::string((char*)img, size);
 	}, quality);
 	return out;
@@ -113,7 +113,7 @@ std::string jwp_save_jpeg_gray_mem_as_string(const fs_image_matrix & matrix, con
 std::vector<uint8_t> jwp_save_jpeg_gray_mem_as_vector(const fs_image_matrix & matrix, const unsigned int quality) {
 	std::vector<uint8_t> out;
 	save_jpeg_gray_mem(matrix,
-		[&](const uint8_t *img, unsigned long size) {
+		[&](const uint8_t *img, size_t size) {
 		out = std::vector<uint8_t>(img, img + size);
 	}, quality);
 	return out;
